arrow string-type casting; updating tests
